// src/sync-client/model/changes.ts
/**
 * Anki Sync Api
 * Api for Anki Synchronization
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DeckConf } from './deck-conf';
import { AppConfiguration } from './app-configuration';
import { Model } from './model';
import { Deck } from './deck';


export interface Changes { 
    models: Array<Model>;
    decks: Array<Array<Deck | DeckConf>>;
    tags: Array<string>;
    conf?: AppConfiguration;
    crt?: number;
}