// src/sync-client/model/deck.ts
/**
 * Anki Sync Api
 * Api for Anki Synchronization
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Deck { 
    newToday: Array<Array<number>>;
    revToday: Array<Array<number>>;
    lrnToday: Array<Array<number>>;
    timeToday: Array<Array<number>>;
    conf: number;
    usn: number;
    desc: string;
    dyn: number;
    collapsed: boolean;
    browserCollapsed: boolean;
    extendNew: number;
    extendRev: number;
    name: string;
    id: number;
    mod: number;
}