// src/sync-client/model/deck-conf.ts
/**
 * Anki Sync Api
 * Api for Anki Synchronization
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ReviewCardConf } from './review-card-conf';
import { LapseCardConf } from './lapse-card-conf';
import { NewCardConf } from './new-card-conf';


export interface DeckConf { 
    id: number;
    mod: number;
    name: string;
    usn: number;
    maxTaken: number;
    autoplay: boolean;
    timer: number;
    replayq: boolean;
    _new: NewCardConf;
    rev: ReviewCardConf;
    lapse: LapseCardConf;
    dyn: number;
    newMix: number;
    newPerDayMinimum: number;
    interdayLearningMix: number;
    reviewOrder: number;
}